added routes, added toggle edit form
