+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter, Switch } from 'react-router-dom';
+import { BrowserRouter} from 'react-router-dom';
 import './index.css';
 import App from './App';
 // import 'bootstrap/dist/css/bootstrap.min.css';
@@ -9,9 +9,7 @@ import reportWebVitals from './reportWebVitals';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <BrowserRouter>
-    <Switch>
      <App />
-    </Switch>
   </BrowserRouter>
 );
